+++ examples/getProjects.ts
@@ -4,7 +4,7 @@ import { generateWalletConnection } from './libs/walletConnection';
 (async () => {
   try {
     const client = new ImmutableX(Config.SANDBOX);
-    const wallet = await generateWalletConnection('goerli');
+    const wallet = await generateWalletConnection('sepolia');
     const pageSize = 200;
     const projects = await client.getProjects(wallet.ethSigner, pageSize);