+++ examples/libs/walletConnection.ts
@@ -1,4 +1,4 @@
-import { AlchemyProvider } from '@ethersproject/providers';
+import { EtherscanProvider } from '@ethersproject/providers';
 import { Wallet } from '@ethersproject/wallet';
 import { createStarkSigner, WalletConnection } from '@imtbl/core-sdk';
 import { requireEnvironmentVariable } from './utils';
@@ -11,10 +11,10 @@ export const generateWalletConnection = async (
 ): Promise<WalletConnection> => {
   const userPrivateKey = requireEnvironmentVariable('PRIVATE_KEY');
   const userStarkKey = requireEnvironmentVariable('STARK_PRIVATE_KEY');
-  const alchemyKey = requireEnvironmentVariable('ALCHEMY_API_KEY');
+  const etherscanKey = requireEnvironmentVariable('ETHERSCAN_API_KEY');
 
   // connect provider
-  const provider = new AlchemyProvider(ethNetwork, alchemyKey);
+  const provider = new EtherscanProvider(ethNetwork, etherscanKey);
 
   // L1 credentials
   const ethSigner = new Wallet(userPrivateKey).connect(provider);