+++ src/config/config.ts
@@ -84,9 +84,9 @@ export const Config = {
   get SANDBOX() {
     return createConfig({
       basePath: 'https://api.sandbox.x.immutable.com',
-      chainID: 5,
-      coreContractAddress: '0x7917eDb51ecD6CdB3F9854c3cc593F33de10c623',
-      registrationContractAddress: '0x1C97Ada273C9A52253f463042f29117090Cd7D83',
+      chainID: 11155111,
+      coreContractAddress: '0x2d5C349fD8464DA06a3f90b4B0E9195F3d1b7F98',
+      registrationContractAddress: '0xDbA6129C02E69405622fAdc3d5A7f8d23eac3b97',
     });
   },